+++ src/index.css
@@ -38,11 +38,11 @@ body {
   flex-direction: column;
 }
 
-.w70{
+.w70 {
   width: 70%;
 }
 
-.w60{
+.w60 {
   width: 60%;
 }
 
@@ -145,7 +145,7 @@ body {
 }
 
 .pad3 {
-  padding: 6em 3em ;
+  padding: 6em 3em;
 }
 
 .gb_gray {
@@ -159,7 +159,7 @@ body {
   align-items: center;
   padding: 3em;
   border-radius: 16px;
-  background-color: #ffffff;  
+  background-color: #fff;
 }
 
 .form {
@@ -184,7 +184,7 @@ body {
   max-width: 50%;
   border: 1px solid #93d14c;
   background-color: #93d14c;
-  color: #ffffff;
+  color: #fff;
   font-size: 0.9em;
   font-weight: 700;
   cursor: pointer;
@@ -214,7 +214,7 @@ body {
 
 .rotated-icon {
   transform: rotate(180deg);
-  color: #ffffff;
+  color: #fff;
 }
 
 .back-button {
@@ -232,7 +232,7 @@ body {
 
 .row-list li {
   padding: 10px;
-  background-color: #ffffff;
+  background-color: #fff;
 }
 
 .row-list li:nth-child(odd) {
@@ -253,7 +253,6 @@ body {
   border: none;
 }
 
-
 .yellow {
   color: rgb(233, 233, 37);
 }
@@ -262,7 +261,7 @@ body {
   border: 1px solid #93d14c;
   border-radius: 20px;
   background-color: #93d14c;
-  color: #ffffff;
+  color: #fff;
   font-size: 1em;
   font-weight: 700;
   cursor: pointer;
