+++ src/routes/DeleteCourse.jsx
@@ -1,11 +1,11 @@
-import { courses } from './Courses';
+import { courses } from '../components/courseData'
 
 const DeleteCourse = () => {
   return (
     <div>
-      <div class="flex flex_col ai_center table-container pad gap_2">
+      <div className="flex flex_col ai_center table-container pad gap_2">
         <h2 className="title">ALL COURSES</h2>
-        <table class="table">
+        <table className="table">
           <thead>
             <tr>
               <th>Course Name</th>
@@ -14,10 +14,12 @@ const DeleteCourse = () => {
           </thead>
           <tbody>
           {courses.map((course) => (
-            <tr>
+            <tr key={course.id}>
               <td>{course.name}</td>
               <td>
-                <button className="form-submit" type="button">Delete</button>
+                <button className="form-submit" type="button">
+                  Delete
+                </button>
               </td>
             </tr>
           ))}
